layer.conf: Set a more generic tune for ARMv8 SoCs
Set a more generic tune for our i.MX8 and i.MX9 machines. This is
particularly needed for the onnxruntime package to compile, as it
requires FP16 instructions. The machines from meta-phytec use a narrow
set with e.g. cortex-a53 tune flags, which in this case excludes
floating-point instructions. As a workaround, include the DEFAULTTUNE
settings in the local.conf of this layer, until meta-phytec or
meta-ampliphy set these appropriately in a distro.

Signed-off-by: Martin Schwan <m.schwan@phytec.de>