demo-celebrity-face-match: Fix systemd service
Fix the systemd service file, so the demo start properly again. In
particular, XDG_RUNTIME_DIR must be set in order for the graphical
application to start.

Remove some redundant settings in the service file.

Also make use of the command line arguments added for the demo to
configure camera and screen.

Signed-off-by: Martin Schwan <m.schwan@phytec.de>